vendor: bundle the DSFR within the app
Right, so until now we were relying on the host app to provide the
DSFR. That's not great because we provide components that spit out
DSFR markup, which is obviously tied to their current release.

So if DSFR version 1000 says a button is:

<button class="fr-btn">Oui</button>

but the DSFR version 2000 says it's actually:

<button class="fr-button">Oui</button>

We have absolutely no way to adjust. So instead ship some bits of the
DSFR ourselves :

* the minimum stylesheet;
* the icons;
* the fonts.

This also means that the host apps don't need to install NodeJS
anymore if they solely rely on it to install the DSFR. In other words,
if your only package is `@gouvfr/dsfr`, you can now just use the gem
and remove Node from your pipeline. Glory glory hallelujah!

This was inspired by my fellow DSFR packager Sylvain during a coffee
break.

Suggested-by: Sylvain Boissel <sylvain.boissel@beta.gouv.fr>