npm: use the dsfr-assets gem instead of Node
Pros:

- it means we can enforce coherent CSS/HTML: relying on others to
provide the assets is problematic since they are both narrowly linked;
- it removes the need for Node in projects that solely depend on it to
install the @gouvfr/dsfr package;
- projects who are supporting a different version of the DSFR can
opt-in: if you don't reference the assets they won't be compiled, we
don't force them upon you – yet, because eventually it would be nice
to maintain that consistency.

Cons:

- your app's installed size is +1.45meg heavier;
- upgrading dsfr-view-components will sometimes mean you have to
update your DSFR version as well, which is perfectly logical.