alert: refactor to use shared HTML classes spec
I was initially testing the shared class matcher (it_behaves_like 'a
component that accepts custom HTML attributes') because we'll be doing
some work with the HTML classes soon (namely removing the `classes`
param).

I ended up rewriting a fair bit of the spec to avoid duplication, and
some of the main component which I couldn't understand.